Issue #150 SidedProxy Magic, bumped some dependency requirements.
