+++ build.gradle
@@ -33,7 +33,7 @@ plugins {
 apply plugin: "net.minecraftforge.gradle.forge"
 apply plugin: "maven"
 
-version = "1.12-0.3.4" + (System.getenv("BUILD_NUMBER") != null ? "." + System.getenv("BUILD_NUMBER") : "")
+version = "1.12-0.3.5" + (System.getenv("BUILD_NUMBER") != null ? "." + System.getenv("BUILD_NUMBER") : "")
 def mod_version = version
 group= "exnihilocreatio" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "exnihilocreatio"
@@ -97,8 +97,8 @@ dependencies
 {
     deobfCompile "mezz.jei:jei_1.12.2:4.11.0.206"
 
-    deobfCompile "slimeknights.mantle:Mantle:1.12-1.3.2.28"
-    deobfCompile "slimeknights:TConstruct:1.12.2-2.10.1.96"
+    deobfCompile "slimeknights.mantle:Mantle:1.12-1.3.3.39"
+    deobfCompile "slimeknights:TConstruct:1.12.2-2.12.0.115"
 
     compileOnly "mcp.mobius.waila:Hwyla:1.8.26-B41_1.12.2:api"
     runtime "mcp.mobius.waila:Hwyla:1.8.26-B41_1.12.2"
@@ -115,7 +115,7 @@ dependencies
 
     compileOnly "org.projectlombok:lombok:1.18.0"
 
-    compile group: "net.shadowfacts", name: "Forgelin", version: "1.7.3"
+    compile group: "net.shadowfacts", name: "Forgelin", version: "1.8.2"
     compile("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
 }
